+++ src/Calc-ChemMassPercent.py
@@ -9,7 +9,7 @@ import io, sys, os
 import re
 import argparse
 
-_version = "Wed Jan  3 04:47:04 TST 2024"
+_version = "Sun Jan  7 19:00:59 TST 2024"
 _code    = "MyCommands(LINUX+WINDOWS/PYTHON3/UTF-8)"
 
 ## switch stdio by platform
@@ -112,6 +112,35 @@ def get_args():
             <Note>
             In this case, if the total volume of aqueous solution
             in the formula is zero, a "div/0" error will be returned.
+
+        Molarity mode (-m, --molar) syntax:
+            Required: liquid volume unit, molecule name
+            Optional: density
+            
+            1. With volume unit and molecule name:
+            -> "1000 ml : 35% HCl" | python Calc-ChemMassPercent.py --molar
+
+                Type          : Solution.1
+                Formula       : 1000ml:35%HCl
+                Volume        : 1000 ml
+                HCl           : 350.0 g / 36.461 amu / 1.0 L = 9.599 mol/L (M)
+            
+            2. With density (density * volume):
+            -> "1.17 * 1000 g: 35% HCl" | python Calc-ChemMassPercent.py --molar
+
+                Type          : Solution.1
+                Formula       : 1.17*1000g:35%HCl
+                Volume        : 1170.0 g (1.000 L)
+                HCl           : 409.5 g / 36.461 amu / 1.0 L = 11.231 mol/L (M)
+
+            <Note>
+            - If the solvent unit is liquid volume rather than mass, 
+              the concentration of the solvent will be w/v% instead of w/w%.
+                - 1st examples unit: solution = "ml", solvent = "w/v %"
+                - 2nd examples unit: solution = "g",  solvent = "w/w %"
+            - By specifying the density, the molarity (mol/L) can be calculated.
+            - Density must be written to the left of the liquid volume.
+            - When specifying density, the liquid volume unit is mass.
         
         Practice:
             <Q.1>
@@ -137,6 +166,19 @@ def get_args():
                 express mixtures of Solutions:
 
                 100 : 10%, 3% = 100 : 10/100, 3/100*1.0
+
+            <Q.3>
+            Calculate the molar concentration (mol/L)
+            of CH3COOH
+            with a density of 1.05 g/mL at 20℃
+            and a purity of 100 w/w%.
+
+                "1.05 * 1000 g : 100% CH3COOH"| python Calc-ChemMassPercent.py --molar
+
+                Type          : Solution.1
+                Formula       : 1.05*1000g:100%CH3COOH
+                Volume        : 1050.0 g (1.000 L)
+                CH3COOH       : 1050.0 g / 60.052 amu / 1.0 L = 17.485 mol/L (M)
 
     Thanks:
         MathPython
@@ -302,6 +344,7 @@ def get_args():
     #parser.print_help()
     ts = lambda x:list(map(str, x.split(';')))
     parser.add_argument("-f", "--formula", help="molecular formula", type=ts)
+    parser.add_argument("-m", "--molar", help="calc molar concentration", action="store_true")
     parser.add_argument("-r", "--round", help="round", default="3", type=str)
     parser.add_argument("-v", "--verbose", help="verbose output", action="store_true")
     parser.add_argument("-d", "--debug", help="debug", action="store_true")
@@ -337,41 +380,84 @@ def splitTermToNumAndUnit(term, compiled_regex):
     str_num  = compiled_regex.sub(r'\1', term)
     str_num  = str_num.replace(',', '').replace('_', '').rstrip('-/*')
     str_unit = compiled_regex.sub(r'\2', term)
+    split_density = None
+    split_vol = None
     if re.search(r'\*|\/', str_num):
+        # split d, vol
+        split_num = str_num.lstrip(r'*').rstrip(r'*').split(r'*')
+        if len(split_num) > 1:
+            split_density = split_num[0].strip()
+            split_vol = split_num[-1].strip()
         str_num = str(eval(str_num))
     if str_unit == '':
         pass
     elif not isNumber(str_num):
         raise_error("Could not get Volume from '{}'.".format(term))
         sys.exit(1)
-    return str_num, str_unit
+    return str_num, str_unit, split_density, split_vol
 
 def getSolventUnit(unit):
     if unit == '':
-        ret = r''
+        ret = [unit, None]
     elif unit.lower() == r'kL'.lower():
-        ret = r't'
+        ret = [r't', float(1*1000*1000)]
     elif unit.lower() == r'L'.lower():
-        ret = r'kg'
+        ret = [r'kg', float(1*1000)]
     elif unit.lower() == r'mL'.lower():
-        ret = r'g'
+        ret = [r'g', float(1*1)]
+    elif unit.lower() == r't'.lower():
+        ret = [unit, float(1*1000*1000)]
     elif unit.lower() == r'kg'.lower():
-        ret = unit
+        ret = [unit, float(1*1000)]
     elif unit.lower() == r'g'.lower():
-        ret = unit
+        ret = [unit, float(1*1)]
     elif unit.lower() == r'mg'.lower():
-        ret = unit
+        ret = [unit, float(1/1000)]
     else:
-        ret = unit
+        ret = [unit, float(1*1)]
+    return ret
+
+def getSolutionUnit(unit):
+    # ret = [str("unit"), float(Volume/Weight), str("Volume/Weight")]
+    # convert to g/ml
+    if unit == '':
+        ret = [unit, None, None]
+    elif unit.lower() == r'kL'.lower():
+        ret = [unit, float(1*1000), r'Volume']
+    elif unit.lower() == r'L'.lower():
+        ret = [unit, float(1), r'Volume']
+    elif unit.lower() == r'mL'.lower():
+        ret = [unit, float(1/1000), r'Volume']
+    elif unit.lower() == r't'.lower():
+        ret = [unit, float(1*1000), r'Weight']
+    elif unit.lower() == r'kg'.lower():
+        ret = [unit, float(1*1), r'Weight']
+    elif unit.lower() == r'g'.lower():
+        ret = [unit, float(1/1000), r'Weight']
+    elif unit.lower() == r'mg'.lower():
+        ret = [unit, float(1/1000/1000), r'Weight']
+    else:
+        ret = [unit, float(1*1), None]
     return ret
 
 class TotalVolume:
 
-    def __init__(self, solute_volume = 0, unit = None, name = None):
+    def __init__(self, solute_volume = 0, unit = None, name = None, split_volume = None):
         self.name = name
         self.unit = unit
         self.solute_volume = solute_volume
+        self.split_volume = split_volume
+        if split_volume != None:
+            self.is_split_volume = True
+        else:
+            self.is_split_volume = False
     
+    def isSplitVolume(self):
+        return self.is_split_volume
+
+    def resetIsSplitVolume(self, flag):
+        self.is_split_volume = flag
+
     def sumSoluteVolume(self, add_volume, unit):
         # test unit
         if self.testUnit(unit):
@@ -379,8 +465,17 @@ class TotalVolume:
         else:
             raise_error("Deferent unit detected. '{}' and '{}'".format(unit, self.unit))
 
+    def sumSplitVolume(self, add_volume, unit):
+        if self.testUnit(unit):
+            self.split_volume += add_volume
+        else:
+            raise_error("Deferent unit detected. '{}' and '{}'".format(unit, self.unit))
+
     def getSoluteVolume(self):
         return self.solute_volume
+
+    def getSplitVolume(self):
+        return self.split_volume
 
     def testUnit(self, unit):
         if unit == None or unit == '':
@@ -392,11 +487,16 @@ class TotalVolume:
 
 class Solution:
 
-    def __init__(self, percent, solute_weight, name = None):
+    def __init__(self, percent, solute_weight, name = None, calc_molmass = False):
         self.name = name
         self.percent = percent
         self.solute_weight = solute_weight
         self.solvent_weight = self.initSolventWeight()
+        if calc_molmass:
+            comp = Composition(name)
+            self.molmass = comp.weight
+        else:
+            self.molmass = None
 
     def initSolventWeight(self):
         if self.solute_weight == 0:
@@ -406,6 +506,12 @@ class Solution:
             
         return ret
         
+    def getMolMass(self):
+        if self.molmass == None:
+            return None
+        else:
+            return float(self.molmass)
+    
     def getSoluteWeight(self):
         ret = self.solute_weight - self.getSolventWeight()
         return ret
@@ -447,6 +553,10 @@ class Solid:
 if __name__ == '__main__':
     # get args
     args = get_args()
+
+    # read module
+    if args.molar:
+        from pymatgen.core import Composition
 
     # read formula
     formulas = []
@@ -496,9 +606,6 @@ if __name__ == '__main__':
             solution = solution.strip()
             solution_counter += 1
             solution_id = solution_symbol + str(solution_counter)
-            if True:
-                print("{} : {}".format("Type".ljust(debug_ljust), solution_id))
-                print("{} : {}".format("Formula".ljust(debug_ljust), solution))
             # get Volume and mass percent concentration
             solution = solution.replace(' ', '')
             if re.search(r':', solution):
@@ -513,24 +620,48 @@ if __name__ == '__main__':
                 vol  = str(solution).strip()
                 conc = str('')
             # set solution volume and unit
-            str_vol, str_vol_unit = splitTermToNumAndUnit(vol, regex_separate_num_and_unit)
+            str_vol, str_vol_unit, split_density, split_vol = splitTermToNumAndUnit(vol, regex_separate_num_and_unit)
+            if str_vol_unit == "" and args.molar:
+                raise_error("Please specify solution unit (e.g. L, mL, g, mg, kg) '{}'.".format(vol))
             if solution_counter == 1:
                 # init volume
-                total_volume = TotalVolume(float(str_vol), str_vol_unit)
+                if split_vol != None:
+                    total_volume = TotalVolume(float(str_vol), str_vol_unit, "Total_Volume", float(split_vol))
+                    #print(total_volume.isSplitVolume())
+                    #print(total_volume.getSplitVolume())
+                else:
+                    total_volume = TotalVolume(float(str_vol), str_vol_unit)
                 if str_vol_unit == "":
                     print_solution_unit = r''
                     print_solvent_unit  = r''
+                    isVolume = None
                 else:
-                    print_solution_unit = str_vol_unit
-                    print_solvent_unit  = getSolventUnit(str_vol_unit)
+                    print_solution_unit, coef_solu_to_liter, isVolume = getSolutionUnit(str_vol_unit)
+                    print_solvent_unit, coef_solv_to_gram = getSolventUnit(str_vol_unit)
             else:
                 # add volume
                 total_volume.sumSoluteVolume(float(str_vol), str_vol_unit)
+                if isVolume == "Weight" and total_volume.isSplitVolume():
+                    if split_vol == None:
+                        # reset 
+                        total_volume.resetIsSplitVolume(False)
+                    else:
+                        total_volume.sumSplitVolume(float(split_vol), str_vol_unit)
+                        print(total_volume.getSplitVolume())
             if True:
+                print("{} : {}".format("Type".ljust(debug_ljust), solution_id))
+                print("{} : {}".format("Formula".ljust(debug_ljust), solution))
                 if print_solution_unit == '':
                     print("{} : {}".format("Volume".ljust(debug_ljust), str_vol))
                 else:
-                    print("{} : {} {}".format("Volume".ljust(debug_ljust), str_vol, str_vol_unit))
+                    if isVolume == "Weight" and total_volume.isSplitVolume():
+                        if split_vol != None:
+                            str_split_vol = str("{:." + args.round + "f}").format(float(split_vol) * coef_solu_to_liter)
+                            str_split_vol_unit = 'L'
+                        print("{} : {} {} ({} {})".format("Volume".ljust(debug_ljust), str_vol, str_vol_unit, str_split_vol, str_split_vol_unit))
+                    else:
+                        print("{} : {} {}".format("Volume".ljust(debug_ljust), str_vol, str_vol_unit))
+            
             # get each mass percent concentration nums
             if conc == '':
                 # no conc terms (Volume only)
@@ -542,15 +673,17 @@ if __name__ == '__main__':
             # set solvent name and concentration
             term_counter = 0
             for cterm in conc_terms:
-                # interpret % symbol
                 cterm = cterm.strip()
                 if cterm == '':
                     continue
                 term_counter += 1
+                # interpret % symbol
                 if regex_allowed_num_and_percent.search(cterm):
                     cterm = regex_allowed_num_and_percent.sub(r'\1/100\2', cterm)
-                str_solv_num, str_solv_name = splitTermToNumAndUnit(cterm, regex_separate_num_and_unit)
-                if str_solv_name == '':
+                str_solv_num, str_solv_name, _, _ = splitTermToNumAndUnit(cterm, regex_separate_num_and_unit)
+                if str_solv_name == '' and args.molar:
+                    raise_error("Please specify molecule name '{}'.".format(solution))
+                elif str_solv_name == r'':
                     str_solv_name = term_symbol + str(term_counter)
                 # add to dict
                 if str_solv_name in solvent_dict.keys():
@@ -558,56 +691,113 @@ if __name__ == '__main__':
                     solvent_dict[str_solv_name].sumSolventWeight(float(str_solv_num), float(str_vol))
                 else:
                     # init solvent dict
-                    solvent_dict[str_solv_name] = Solution(float(str_solv_num), float(str_vol), str(str_solv_name))
+                    if args.molar:
+                        solvent_dict[str_solv_name] = Solution(float(str_solv_num), float(str_vol), str(str_solv_name), True)
+                        #print(solvent_dict[str_solv_name].getMolMass())
+                    else:
+                        solvent_dict[str_solv_name] = Solution(float(str_solv_num), float(str_vol), str(str_solv_name))
                 if True:
                     solvent_dict[solid_name].sumSolventWeight(float(str_solv_num), float(str_vol))
                 # output each solvent mass percent
-                if True:
+                if args.molar:
+                    c = str_solv_name.ljust(debug_ljust)
+                    v = str_vol
+                    if isVolume == r"Volume":
+                        vol_unit = r" mol/L (M)"
+                        v_liter = float(v) * coef_solu_to_liter
+                        v_str = str("{} L").format(v_liter)
+                    elif isVolume == r"Weight":
+                        if total_volume.isSplitVolume() and split_vol != None:
+                            vol_unit = r" mol/L (M)"
+                            v_liter = float(split_vol) * coef_solu_to_liter
+                            v_str = str("{} L").format(v_liter)
+                        else:
+                            vol_unit = r" mol/kg"
+                            v_liter = float(v) * coef_solu_to_liter
+                            v_str = str("{} kg").format(v_liter)
+                    else:
+                        v_str = str("{}").format(v_liter)
+                        vol_unit = r''
+                    n = float(str_solv_num)
+                    molmass = float(solvent_dict[str_solv_name].getMolMass())
+                    if float(v) == 0:
+                        w = n
+                        w_gram = w * coef_solv_to_gram
+                        molar = w_gram / molmass
+                    else:
+                        w = n * float(v)
+                        w_gram = w * coef_solv_to_gram
+                        molar = w_gram / molmass / v_liter
+                    w_str = str("{} g").format(w_gram)
+                    m_str = str("{:." + args.round + "f} amu").format(molmass)
+                    p_str = str("{:." + args.round + "f}{}").format(molar, vol_unit)
+                    if print_solution_unit != r'':
+                        u_solv = print_solvent_unit
+                        if float(v) == 0:
+                            # case Volume == 0 (only solvent)
+                            print(str("{} : {} ({} {})").format(c, w_str, w, u_solv))
+                        else:
+                            print(str("{} : {} / {} / {} = {}").format(c, w_str, m_str, v_str, p_str))
+                else:
                     c = str_solv_name.ljust(debug_ljust)
                     v = str_vol
                     n = float(str_solv_num)
-                    p = float(str_solv_num) * 100
-                    w = n * float(v)
-                    n = str("{:." + args.round + "f}").format(n)
-                    p = str("{:." + args.round + "f}").format(p)
+                    p = n * 100
+                    if float(v) == 0:
+                        w = n
+                    else:
+                        w = n * float(v)
+                    if isVolume == r"Volume":
+                        p_unit = "w/v%"
+                    elif isVolume == r"Weight":
+                        p_unit = "w/w%"
+                    else:
+                        p_unit = "%"
+                    n_str = str("{:." + args.round + "f}").format(n)
+                    p_str = str("{:." + args.round + "f} {}").format(p, p_unit)
                     if print_solution_unit == '':
                         if float(v) == 0:
                             # case Volume == 0 (only solvent)
-                            w = n * 1
                             print(str("{} : {}").format(c, w))
                         else:
-                            print(str("{} : {} / {} = {} ({} %)").format(c, w, v, n, p))
+                            print(str("{} : {} / {} = {} ({})").format(c, w, v, n_str, p_str))
                     else:
                         u_solv = print_solvent_unit
                         u_solu = print_solution_unit
                         if float(v) == 0:
                             # case Volume == 0 (only solvent)
-                            w = n * 1
                             print(str("{} : {} {}").format(c, w, u_solv))
                         else:
-                            print(str("{} : {} {} / {} {} = {} ({} %)").format(c, w, u_solv, v, u_solu, n, p))
+                            print(str("{} : {} {} / {} {} = {} ({})").format(c, w, u_solv, v, u_solu, n_str, p_str))
             
             if args.debug or args.verbose:
-                if print_solution_unit == '':
-                    print("{} : {}".format("Total_Volume".ljust(debug_ljust), total_volume.getSoluteVolume()))
-                else:
-                    print("{} : {} {}".format("Total_Volume".ljust(debug_ljust), total_volume.getSoluteVolume(), print_solution_unit))
-                if len(solvent_dict) > 0:
-                    for key in solvent_dict.keys():
-                        c = str("Total_" + key).ljust(debug_ljust)
-                        v = total_volume.getSoluteVolume()
-                        w = solvent_dict[key].solvent_weight
-                        r = w / v
-                        p = w / v * 100
-                        r = str("{:." + args.round + "f}").format(r)
-                        p = str("{:." + args.round + "f}").format(p)
-                        if print_solution_unit == '':
-                            print(str("{} : {} / {} = {} ({} %)").format(c, w, v, r, p))
-                        else:
-                            u_solv = print_solvent_unit
-                            u_solu = print_solution_unit
-                            print(str("{} : {} {} / {} {} = {} ({} %)").format(c, w, u_solv, v, u_solu, r, p))
-            
+                # output 
+                if not args.molar:
+                    if print_solution_unit == '':
+                        print("{} : {}".format("Total_Volume".ljust(debug_ljust), total_volume.getSoluteVolume()))
+                    else:
+                        print("{} : {} {}".format("Total_Volume".ljust(debug_ljust), total_volume.getSoluteVolume(), print_solution_unit))
+                    if len(solvent_dict) > 0:
+                        for key in solvent_dict.keys():
+                            c = str("Total_" + key).ljust(debug_ljust)
+                            v = total_volume.getSoluteVolume()
+                            w = solvent_dict[key].solvent_weight
+                            r = w / v
+                            p = w / v * 100
+                            if isVolume == r"Volume":
+                                p_unit = "w/v%"
+                            elif isVolume == r"Weight":
+                                p_unit = "w/w%"
+                            else:
+                                p_unit = "%"
+                            r_str = str("{:." + args.round + "f}").format(r)
+                            p_str = str("{:." + args.round + "f} {}").format(p, p_unit)
+                            if print_solution_unit == '':
+                                print(str("{} : {} / {} = {} ({})").format(c, w, v, r_str, p_str))
+                            else:
+                                u_solv = print_solvent_unit
+                                u_solu = print_solution_unit
+                                print(str("{} : {} {} / {} {} = {} ({})").format(c, w, u_solv, v, u_solu, r_str, p_str))            
             if True:
                 print("")
         
@@ -620,36 +810,95 @@ if __name__ == '__main__':
             print("{} : {}".format(n, v))
         else:
             u = print_solution_unit
-            print("{} : {} {}".format("Total_Volume".ljust(debug_ljust), total_volume.getSoluteVolume(), u))
+            if isVolume == "Weight" and total_volume.isSplitVolume():
+                vs = total_volume.getSplitVolume()
+                ts = str("{:." + args.round + "f}").format(vs * coef_solu_to_liter)
+                print("{} : {} {} ({} L)".format("Total_Volume".ljust(debug_ljust), v, u, ts))
+            else:
+                print("{} : {} {}".format("Total_Volume".ljust(debug_ljust), v, u))
         if len(solvent_dict) > 0:
             for key in solvent_dict.keys():
-                if key != solid_name:
+                if args.molar:
+                    if key == solid_name:
+                        # do not parse total solid
+                        continue
+                    c = str("Total_" + key).ljust(debug_ljust)
+                    if isVolume == r"Volume":
+                        vol_unit = r" mol/L (M)"
+                        v_liter = float(v) * coef_solu_to_liter
+                        v_str = str("{} L").format(v_liter)
+                    elif isVolume == r"Weight":
+                        if total_volume.isSplitVolume():
+                            vol_unit = r" mol/L (M)"
+                            v_liter = vs * coef_solu_to_liter
+                            v_str = str("{} L").format(v_liter)
+                        else:
+                            # split_vol == None:
+                            vol_unit = r" mol/kg"
+                            v_liter = float(v) * coef_solu_to_liter
+                            v_str = str("{} kg").format(v_liter)
+                    else:
+                        v_str = str("{}").format(v_liter)
+                        v_liter = float(v) * coef_solu_to_liter
+                        vol_unit = r""
+                    w = solvent_dict[key].solvent_weight
+                    w_gram = w * coef_solv_to_gram
+                    molmass = float(solvent_dict[key].getMolMass())
+                    if float(v) == 0:
+                        molar = w_gram / molmass
+                    else:
+                        molar = w_gram / molmass / v_liter
+                    w_str = str("{} g").format(w_gram)
+                    m_str = str("{:." + args.round + "f} amu").format(molmass)
+                    p_str = str("{:." + args.round + "f}{}").format(molar, vol_unit)
+                    if print_solution_unit != r'':
+                        u_solv = print_solvent_unit
+                        if float(v) == 0:
+                            # case Volume == 0 (only solvent)
+                            print(str("{} : {} ({} {})").format(c, w_str, w, u_solv))
+                        else:
+                            print(str("{} : {} / {} / {} = {}").format(c, w_str, m_str, v_str, p_str))
+                elif key != solid_name:
                     c = str("Total_" + key).ljust(debug_ljust)
                     v = total_volume.getSoluteVolume()
                     w = solvent_dict[key].solvent_weight
                     r = w / v
                     p = w / v * 100
-                    r = str("{:." + args.round + "f}").format(r)
-                    p = str("{:." + args.round + "f}").format(p)
+                    if isVolume == r"Volume":
+                        p_unit = "w/v%"
+                    elif isVolume == r"Weight":
+                        p_unit = "w/w%"
+                    else:
+                        p_unit = "%"
+                    r_str = str("{:." + args.round + "f}").format(r)
+                    p_str = str("{:." + args.round + "f} {}").format(p, p_unit)
                     if print_solution_unit == '':
-                        print(str("{} : {} / {} = {} ({} %)").format(c, w, v, r, p))
+                        print(str("{} : {} / {} = {} ({})").format(c, w, v, r_str, p_str))
                     else:
                         u_solv = print_solvent_unit
                         u_solu = print_solution_unit
-                        print(str("{} : {} {} / {} {} = {} ({} %)").format(c, w, u_solv, v, u_solu, r, p))
-            key = solid_name
-            c = str("Total_" + key).ljust(debug_ljust)
-            v = total_volume.getSoluteVolume()
-            w = solvent_dict[key].solvent_weight
-            r = w / v
-            p = w / v * 100
-            r = str("{:." + args.round + "f}").format(r)
-            p = str("{:." + args.round + "f}").format(p)
-            if print_solution_unit == '':
-                print(str("{} : {} / {} = {} ({} %)").format(c, w, v, r, p))
-            else:
-                u_solv = print_solvent_unit
-                u_solu = print_solution_unit
-                print(str("{} : {} {} / {} {} = {} ({} %)").format(c, w, u_solv, v, u_solu, r, p))
+                        print(str("{} : {} {} / {} {} = {} ({})").format(c, w, u_solv, v, u_solu, r_str, p_str))
+            if not args.molar:
+                # output total solid mass percent
+                key = solid_name
+                c = str("Total_" + key).ljust(debug_ljust)
+                v = total_volume.getSoluteVolume()
+                w = solvent_dict[key].solvent_weight
+                r = w / v
+                p = w / v * 100
+                if isVolume == r"Volume":
+                    p_unit = "w/v%"
+                elif isVolume == r"Weight":
+                    p_unit = "w/w%"
+                else:
+                    p_unit = "%"
+                r_str = str("{:." + args.round + "f}").format(r)
+                p_str = str("{:." + args.round + "f} {}").format(p, p_unit)
+                if print_solution_unit == '':
+                    print(str("{} : {} / {} = {} ({})").format(c, w, v, r_str, p_str))
+                else:
+                    u_solv = print_solvent_unit
+                    u_solu = print_solution_unit
+                    print(str("{} : {} {} / {} {} = {} ({})").format(c, w, u_solv, v, u_solu, r_str, p_str))
     
     sys.exit(0)