+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.4.1] - 2024-01-04
 
 - Updated links
+- Added [Calc-ChemMassPercent.py] `--molar` option
 - Updated [Calc-ChemMassPercent.py] synopsis
 
 ## [1.4.0] - 2024-01-04